refactor(store): split atoms.ts into domain modules
Extract layout, UI, and audio atoms into separate domain modules:
- layoutAtoms.ts: fretStart, fretEnd, fretZoom, tuning, currentTuning
- uiAtoms.ts: displayFormat, mobileTab, tabletTab, landscapeNarrowTab
- audioAtoms.ts: enharmonicDisplay, isMuted, toggleMute

Existing modules retained: fingeringAtoms, scaleAtoms, chordOverlayAtoms, practiceLensAtoms

atoms.ts now re-exports all domains for backward compatibility.
